backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=1.30.0

pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
motor==3.3.2
python-multipart==0.0.9
